Separate HasList test
