+++ tests/Unit/Library/Fields/SelectFieldTest.php
@@ -37,25 +37,4 @@ class SelectFieldTest extends TestCase
     {
         $this->assertMatchesHtmlSnapshot($this->field->renderInput());
     }
-
-    /** @test */
-    public function get_list(): void
-    {
-        $this->assertSame(
-            ['item 1', 'item 2', 'item 3'],
-            $this->field->getList(),
-        );
-    }
-
-    /** @test */
-    public function get_list_default(): void
-    {
-        // Create a field instance without a config that contains a `list` key.
-        $field = new SelectField('test_key', 'select');
-
-        $this->assertSame(
-            [],
-            $field->getList(),
-        );
-    }
 }
